+++ core/pom.xml
@@ -513,69 +513,7 @@
   <build>
     <outputDirectory>target/scala-${scala.binary.version}/classes</outputDirectory>
     <testOutputDirectory>target/scala-${scala.binary.version}/test-classes</testOutputDirectory>
-    <resources>
-      <resource>
-        <directory>${project.basedir}/src/main/resources</directory>
-      </resource>
-      <resource>
-        <!-- Include the properties file to provide the build information. -->
-        <directory>${project.build.directory}/extra-resources</directory>
-        <filtering>true</filtering>
-      </resource>
-    </resources>
     <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-antrun-plugin</artifactId>
-        <executions>
-          <execution>
-            <id>choose-shell-and-script</id>
-            <phase>validate</phase>
-            <goals>
-              <goal>run</goal>
-            </goals>
-            <configuration>
-              <exportAntProperties>true</exportAntProperties>
-              <target>
-                <condition property="shell" value="powershell.exe" else="bash">
-                  <and>
-                    <os family="windows"/>
-                  </and>
-                </condition>
-                <condition property="spark-build-info-script" value="spark-build-info.ps1"
-                           else="spark-build-info">
-                  <and>
-                    <os family="windows"/>
-                  </and>
-                </condition>
-                <echo>Shell to use for generating spark-version-info.properties file =
-                  ${shell}
-                </echo>
-                <echo>Script to use for generating spark-version-info.properties file =
-                  ${spark-build-info-script}
-                </echo>
-              </target>
-            </configuration>
-          </execution>
-          <execution>
-            <id>generate-spark-build-info</id>
-            <phase>generate-resources</phase>
-            <configuration>
-              <!-- Execute the shell script to generate the spark build information. -->
-              <target>
-                <exec executable="${shell}">
-                  <arg value="${project.basedir}/../build/${spark-build-info-script}"/>
-                  <arg value="${project.build.directory}/extra-resources"/>
-                  <arg value="${project.version}"/>
-                </exec>
-              </target>
-            </configuration>
-            <goals>
-              <goal>run</goal>
-            </goals>
-          </execution>
-        </executions>
-      </plugin>
       <plugin>
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-dependency-plugin</artifactId>
